revamp flood event handling
This makes a FloodConfig type with a start and duration, so that the
start and end are associated with each other.

In some future commit this will let us do useful things while in a flood,
like monitor the pump power.

Switch to `tasmor` for the Tasmota interface instead of making http
requests of the Tasmota device by hand.